parameters service as lib service
