+++ Interfaces/Mongo/MongoBaseService.cs
@@ -1,7 +1,10 @@
 ﻿using MongoDB.Driver;
+using System;
+using System.Collections.Generic;
 using System.Linq.Expressions;
+using System.Threading.Tasks;
 
-namespace Parameters.API.Services.Mongo
+namespace Interfaces.Mongo
 {
     public abstract class MongoBaseService<T>
     {
