--- src/Services/Parameters.API/Parameters.API/Dockerfile
@@ -1,22 +1,0 @@
-#See https://aka.ms/containerfastmode to understand how Visual Studio uses this Dockerfile to build your images for faster debugging.
-
-FROM mcr.microsoft.com/dotnet/aspnet:6.0 AS base
-WORKDIR /app
-EXPOSE 80
-EXPOSE 443
-
-FROM mcr.microsoft.com/dotnet/sdk:6.0 AS build
-WORKDIR /src
-COPY ["src/Services/Parameters.API/Parameters.API/Parameters.API.csproj", "src/Services/Parameters.API/Parameters.API/"]
-RUN dotnet restore "src/Services/Parameters.API/Parameters.API/Parameters.API.csproj"
-COPY . .
-WORKDIR "/src/src/Services/Parameters.API/Parameters.API"
-RUN dotnet build "Parameters.API.csproj" -c Release -o /app/build
-
-FROM build AS publish
-RUN dotnet publish "Parameters.API.csproj" -c Release -o /app/publish
-
-FROM base AS final
-WORKDIR /app
-COPY --from=publish /app/publish .
-ENTRYPOINT ["dotnet", "Parameters.API.dll"]
