--- src/Services/Parameters.API/Parameters.API/Models/Requests/InsertRequest.cs
@@ -1,9 +1,0 @@
-﻿using Parameters.API.Models.Documents;
-
-namespace Parameters.API.Models.Requests
-{
-    public class InsertRequest
-    {
-        public List<Parameter> Parameters { get; set; }
-    }
-}
