--- src/Services/Parameters.API/Parameters.API/Program.cs
@@ -1,53 +1,0 @@
-using Parameters.API;
-using Parameters.API.Services;
-using Parameters.API.Services.Store;
-
-var builder = WebApplication.CreateBuilder(args);
-
-var services = builder.Services;
-
-// Add services to the container.
-services.AddControllers().AddNewtonsoftJson();
-
-services.AddSwaggerGen(c =>
-{
-    c.SwaggerDoc("v1", new Microsoft.OpenApi.Models.OpenApiInfo() { });
-});
-
-//stores
-services.AddMongoService(builder.Configuration);
-services.AddSingleton<ParametersStore>();
-
-//services
-services.AddSingleton<ParametersService>();
-
-var app = builder.Build();
-
-// Configure the HTTP request pipeline.
-if (!app.Environment.IsDevelopment())
-{
-    app.UseExceptionHandler("/Error");
-    // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
-    app.UseHsts();
-}
-
-app.UseHttpsRedirection();
-app.UseStaticFiles();
-
-app.UseRouting();
-
-app.UseAuthorization();
-
-app.UseSwagger();
-
-app.UseSwaggerUI(c =>
-{
-    c.SwaggerEndpoint("/swagger/v1/swagger.json", "Agents API");
-    c.RoutePrefix = string.Empty;
-});
-
-app.MapControllerRoute(
-    name: "default",
-    pattern: "{controller=Home}/{action=Index}/{id?}");
-
-app.Run();
