--- src/Services/Parameters.API/Parameters.API/appsettings.Development.json
@@ -1,14 +1,0 @@
-{
-  "DetailedErrors": true,
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "MongoDBSettings": {
-    "ConnectionString": "mongodb://localhost:27017",
-    "DatabaseName": "APM",
-    "QueryLogging": true
-  }
-}
